+++ server/src/main/java/calendar/event/dto/CreateEventDTO.java
@@ -8,7 +8,7 @@ public class CreateEventDTO {
     private String location;
     private String description;
     private long date;
-    private float duration;
+    private long duration;
     private boolean recurring;
     private int recursEvery;
     private long recursUntil;
@@ -51,11 +51,11 @@ public class CreateEventDTO {
         this.date = date;
     }
 
-    public float getDuration() {
+    public long getDuration() {
         return duration;
     }
 
-    public void setDuration(float duration) {
+    public void setDuration(long duration) {
         this.duration = duration;
     }
